requirements.txt
numpy
scipy==1.9.3
scikit-learn
matplotlib
pandas
fastapi
uvicorn==0.20.0
gunicorn==19.9.0
httptools
Pillow==9.3.0
tensorflow-cpu==2.11.0
pydantic==1.10.2